requirements.txt
functions-framework==3.*
websocket-client
pillow